secure-file-vault/requirements.txt
python-dotenv
boto3==1.38.4
botocore==1.38.4
jmespath==1.0.1
python-dateutil==2.9.0.post0
s3transfer==0.12.0
six==1.17.0
urllib3==2.4.0